Renaming examples
