--- embedding_python_examples/example9.cpp
@@ -1,11 +1,0 @@
-#define PY_SSIZE_T_CLEAN  /* Make "s#" use Py_ssize_t rather than int. */
-#include <Python.h>
-
-int ok;
-int i, j;
-long k, l;
-const char *s;
-Py_ssize_t size;
-
-ok = PyArg_ParseTuple(args, ""); /* No arguments */
-    /* Python call: f() */
